Update image cache class with more configuration options and docs.
Also, don't cache img if it is the default texture.

+++ ardor3d-core/src/main/java/com/ardor3d/util/HttpImageCache.java
@@ -33,6 +33,7 @@ import java.util.regex.Pattern;
 
 import com.ardor3d.image.Image;
 import com.ardor3d.image.util.ImageLoaderUtil;
+import com.ardor3d.renderer.state.TextureState;
 import com.ardor3d.util.export.binary.BinaryExporter;
 import com.ardor3d.util.resource.ResourceSource;
 import com.ardor3d.util.resource.URLResourceSource;
@@ -42,17 +43,94 @@ public enum HttpImageCache {
 
     Comparator<ImageCacheItem> DateCompare = (a, b) -> b.lastAccessed.compareTo(a.lastAccessed);
     ConcurrentMap<String, ImageCacheItem> MemoryCache = new ConcurrentHashMap<String, ImageCacheItem>();
+    boolean checkModified = false;
+    long maxMemoryCacheSize = 8 * 1024 * 1024;
+    long minModCheckSeconds = 24 * 60 * 60;
 
     File cacheDir = new File(System.getProperty("user.dir") + "/cache/image/");
 
+    /**
+     * @return the directory where we will store copies of downloaded images, in .ABI format. If null, file caching will
+     *         be skipped. Defaults to <code>System.getProperty("user.dir") + "/cache/image/"</code>
+     */
     public File getCacheDir() {
         return cacheDir;
     }
 
+    /**
+     * @param cacheDir
+     *            the directory where we want to store copies of downloaded images, in .ABI format. If null, file
+     *            caching will be skipped.
+     */
     public void setCacheDir(final File cacheDir) {
         this.cacheDir = cacheDir;
     }
 
+    /**
+     * @return true if we should ask if a file we have cached locally has been recently modified. This action is taken
+     *         if only if we find we already have the image and have not check in {@link #minModCheckSeconds} seconds.
+     *         Defaults to false.
+     */
+    public boolean isCheckModified() {
+        return checkModified;
+    }
+
+    /**
+     * @param checkModified
+     *            true if we should ask if a file we have cached locally has been recently modified.
+     */
+    public void setCheckModified(final boolean checkModified) {
+        this.checkModified = checkModified;
+    }
+
+    /**
+     * @return the maximum total size, in bytes, of the Images we will hold in our memory cache. When we hit this limit,
+     *         we will flush items out of memory based on LRU logic. Default is 8 MB (8 * 1024 * 1024)
+     */
+    public long getMaxMemoryCacheSize() {
+        return maxMemoryCacheSize;
+    }
+
+    /**
+     * @param maxMemoryCacheSize
+     *            the maximum total size, in bytes, of the Images we will hold in our memory cache. When we hit this
+     *            limit, we will flush items out of memory based on LRU logic.
+     */
+    public void setMaxMemoryCacheSize(final long maxMemoryCacheSize) {
+        this.maxMemoryCacheSize = maxMemoryCacheSize;
+    }
+
+    /**
+     * @return the minimum time, in seconds, between checking an image for remote modification. Default is 1 day (24 *
+     *         60 * 60).
+     */
+    public long getMinModCheckSeconds() {
+        return minModCheckSeconds;
+    }
+
+    /**
+     * @param minModCheckSeconds
+     *            the minimum time, in seconds, between checking an image for remote modification.
+     */
+    public void setMinModCheckSeconds(final long minModCheckSeconds) {
+        this.minModCheckSeconds = minModCheckSeconds;
+    }
+
+    /**
+     * Attempt to retrieve an image, first from the local cache, then from the given URI. Runs asynchronously, calling
+     * the supplied Consumer callback on completion.
+     *
+     * @param uri
+     *            location of the image to download.
+     * @param type
+     *            the extension type of the image (e.g. ".png"). Used by ImageLoaderUtil when loading the image contents
+     *            from the URI.
+     * @param flipped
+     *            if true, flip the image vertically before returning.
+     * @param onSuccess
+     *            Consumer callback called once we've downloaded the image. May be called with the argument
+     *            <code>TextureState.getDefaultTextureImage()</code> if we are unable to load the image.
+     */
     public void loadFromCacheOrDownloadAsync(final URI uri, final String type, final boolean flipped,
             final Consumer<Image> onSuccess) {
         CompletableFuture.runAsync(() -> {
@@ -61,6 +139,19 @@ public enum HttpImageCache {
         });
     }
 
+    /**
+     * Attempt to retrieve an image, first from the local cache, then from the given URI.
+     *
+     * @param uri
+     *            location of the image to download.
+     * @param type
+     *            the extension type of the image (e.g. ".png"). Used by ImageLoaderUtil when loading the image contents
+     *            from the URI.
+     * @param flipped
+     *            if true, flip the image vertically before returning.
+     * @return The loaded Image, or <code>TextureState.getDefaultTextureImage()</code> if we are unable to load the
+     *         image.
+     */
     public Image loadFromCacheOrDownload(final URI uri, final String type, final boolean flipped) {
         // Convert our name to a key to be used in our in memory hash and file system
         final String key = convertUrlToFileName(uri.toString(), type, flipped);
@@ -68,7 +159,7 @@ public enum HttpImageCache {
         // Check the memory cache
         ImageCacheItem cacheItem = MemoryCache.get(key);
 
-        if (cacheItem == null) {
+        if (cacheItem == null && cacheDir != null) {
             // Check the file cache
             try {
                 final File f = new File(cacheDir, key);
@@ -101,9 +192,8 @@ public enum HttpImageCache {
         }
 
         // if we have not asked about this item for a bit, fire off a check to see if there's an update. But send the
-        // cached copy immediately.
-        if (cacheItem.lastChecked.until(Instant.now(), ChronoUnit.MINUTES) > 15) {
-//            System.out.println("firing check...");
+        // cached copy back immediately.
+        if (checkModified && cacheItem.lastChecked.until(Instant.now(), ChronoUnit.SECONDS) > minModCheckSeconds) {
             final ImageCacheItem item = cacheItem;
             CompletableFuture.runAsync(() -> tryGetImage(item, uri));
         } else {
@@ -111,7 +201,6 @@ public enum HttpImageCache {
         }
 
         // return any existing cached copy.
-//        System.out.println("returning cache item (1).");
         return cacheItem.value;
     }
 
@@ -122,11 +211,17 @@ public enum HttpImageCache {
             // No existing value, so simply load the image and add to memory cache
             try {
                 final URLResourceSource source = new URLResourceSource(uri.toURL(), cacheItem.type);
-                cacheItem.value = ImageLoaderUtil.loadImage(source, cacheItem.flipped);
+                final Image img = ImageLoaderUtil.loadImage(source, cacheItem.flipped);
+                if (img == TextureState.getDefaultTextureImage()) {
+                    return img;
+                }
+
+                cacheItem.value = img;
                 if (source.getLastModifiedValue() != 0L) {
                     cacheItem.lastModified = Instant.ofEpochMilli(source.getLastModifiedValue());
                 }
                 MemoryCache.put(cacheItem.id, cacheItem);
+
             } catch (final MalformedURLException ex) {
                 ex.printStackTrace();
                 return null;
@@ -135,9 +230,9 @@ public enum HttpImageCache {
             // we have a value, so this is a check instead. Let's open a connection and see if it has been modified
             try {
                 final URLConnection connection = uri.toURL().openConnection();
-                connection.setIfModifiedSince(cacheItem.lastModified.toEpochMilli());
-                connection.connect();
                 if (connection instanceof HttpURLConnection) {
+                    connection.setIfModifiedSince(cacheItem.lastModified.toEpochMilli());
+                    connection.connect();
                     final HttpURLConnection httpConnection = (HttpURLConnection) connection;
                     final int responseCode = httpConnection.getResponseCode();
                     if (responseCode == HttpURLConnection.HTTP_NOT_MODIFIED) {
@@ -159,6 +254,10 @@ public enum HttpImageCache {
     }
 
     private void updateFileCache(final ImageCacheItem cacheItem) {
+        if (cacheDir == null) {
+            return;
+        }
+
         try {
             if (!cacheDir.exists()) {
                 cacheDir.mkdir();
@@ -192,24 +291,21 @@ public enum HttpImageCache {
         MemoryCache.clear();
 
         // clear the file cache as well.
-        if (cacheDir.exists()) {
+        if (cacheDir != null && cacheDir.exists()) {
             deleteDirectory(cacheDir);
         }
     }
 
     public static boolean deleteDirectory(final File dir) {
         if (dir.isDirectory()) {
-            final File[] children = dir.listFiles();
-            for (int i = 0; i < children.length; i++) {
-                final boolean success = deleteDirectory(children[i]);
-                if (!success) {
+            for (final File element : dir.listFiles()) {
+                if (!deleteDirectory(element)) {
                     return false;
                 }
             }
         }
 
         // either file or an empty directory
-        System.out.println("removing file or directory : " + dir.getName());
         return dir.delete();
     }
 
@@ -219,13 +315,12 @@ public enum HttpImageCache {
 
         int i = 0;
         final int maxI = sortedList.size();
-        int bytes = 0;
-        final int maxBytes = 5 * 1024 * 1024;
+        long bytes = 0L;
 
         ImageCacheItem item;
-        for (; i < maxI && bytes <= maxBytes; i++) {
+        for (; i < maxI && bytes <= maxMemoryCacheSize; i++) {
             item = sortedList.get(i);
-            if (item.getMemorySize() + bytes > maxBytes) {
+            if (item.getMemorySize() + bytes > maxMemoryCacheSize) {
                 // We want to break out at this point and remove anything pass this point
                 break;
             }
@@ -283,7 +378,7 @@ public enum HttpImageCache {
             final StringBuilder sb = new StringBuilder(key);
             md.update(sb.reverse().toString().getBytes());
             final String part2 = bytesToHex(md.digest());
-            return part1 + part2;
+            return part1 + part2 + ".abi";
         } catch (final NoSuchAlgorithmException ex) {
             String rt = "";
             final Matcher matcher = NameRegex.matcher(urlText);
